fixing policy and all controller

+++ app/Http/Controllers/Api/ServiceCategoryController.php
@@ -7,11 +7,16 @@ use Illuminate\Http\Request;
 use App\Models\ServiceCategory;
 use App\Http\Resources\ServiceCategoryResource; // <-- Tambahkan Http di sini
 use Illuminate\Validation\Rule; // <-- TAMBAHKAN BARIS INI
+use Illuminate\Foundation\Auth\Access\AuthorizesRequests; // <-- 1. TAMBAHKAN INI
 
 class ServiceCategoryController extends Controller
 {
+    use AuthorizesRequests; // <-- 2. TAMBAHKAN INI
     public function index()
     {
+        // Cek izin: apakah user boleh melihat daftar Service Category?
+        $this->authorize('viewAny', ServiceCategory::class);
+
         // Gunakan Resource untuk konsistensi output
         return ServiceCategoryResource::collection(ServiceCategory::all());
     }
@@ -21,6 +26,9 @@ class ServiceCategoryController extends Controller
      */
     public function store(Request $request)
     {
+        // Cek izin: apakah user boleh melihat daftar Service Category?
+        $this->authorize('create', ServiceCategory::class);
+
         // 1. Validasi input
         $validated = $request->validate([
             'name' => 'required|string|unique:service_categories|max:255',
@@ -41,6 +49,9 @@ class ServiceCategoryController extends Controller
      */
     public function show(ServiceCategory $serviceCategory)
     {
+        // Cek izin: apakah user boleh melihat Service Category ini?
+        $this->authorize('view', $serviceCategory);
+
         // Laravel otomatis mencari ServiceCategory berdasarkan ID dari URL.
         // Jika tidak ketemu, otomatis akan menampilkan error 404.
         return new ServiceCategoryResource($serviceCategory);
@@ -51,6 +62,9 @@ class ServiceCategoryController extends Controller
      */
     public function update(Request $request, ServiceCategory $serviceCategory)
     {
+        // Cek izin: apakah user boleh melihat Service Category ini?
+        $this->authorize('update', $serviceCategory);
+
         // Validasi, dengan aturan 'unique' yang mengabaikan ID serviceCategory saat ini
         $validated = $request->validate([
             'name' => [
@@ -72,6 +86,8 @@ class ServiceCategoryController extends Controller
      */
     public function destroy(ServiceCategory $serviceCategory)
     {
+        // Cek izin: apakah user boleh menghapus Service Category ini?
+        $this->authorize('delete', $serviceCategory);
         $serviceCategory->delete();
 
         // Kembalikan response "204 No Content", standar untuk delete yang sukses
